chore(clippy): zero out workspace clippy warnings + lock baseline (#798)
Resolve all 266 workspace clippy warnings and lock the baseline.

- Phase A: cargo clippy --fix mechanical autofixes (~162).
- Phase B: hand-fixed remainder properly (struct-update syntax,
  const{assert!}, RangeInclusive::contains, let-chains, clamp,
  sort_by_key, doc indentation, unused imports, etc.).
- Phase C: structural lints (too_many_arguments, type_complexity,
  large_enum_variant, module_inception) get targeted #[allow] with
  rationale — no signature/type refactors. Genuinely-orphaned dead
  code removed; intentional dead code allowed with rationale.
  cfg(test)-only imports gated with #[cfg(test)] (not deleted).
- Phase D: [workspace.lints.clippy] in root + lints.workspace=true in
  members, so the baseline can't silently regress.
- Dockerfile: re-copy root Cargo.toml after the crates COPY — cargo
  chef cook rewrites it to a recipe skeleton that drops
  [workspace.lints], which members inherit. Caught by the infra smoke.
- patches/anndists: drop one unused import so the standard
  `cargo clippy --workspace -- -D warnings` gate passes without excludes.

No behavioral changes; no real bugs surfaced (the 18 non-binding
futures were intentional detached spawns / a compile-time signature
check). Closes #798.

+++ crates/unimatrix-server/src/eval/runner/replay.rs
@@ -146,7 +146,7 @@ async fn run_single_profile(
         .entries
         .into_iter()
         .map(|se| ScoredEntry {
-            id: se.entry.id as u64,
+            id: se.entry.id,
             title: se.entry.title.clone(),
             content: se.entry.content.clone(),
             category: se.entry.category.clone(),
